# services/adapter-fabric/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TANTU Adapter Fabric",
    "description": "Multi-protocol shopfloor ingestion \u2014 OPC-UA, Modbus, MQTT, MTConnect, EtherNet/IP, Camera \u2014 normalized to one schema, derived DefectEvents only (no image field).",
    "version": "0.1.0"
  },
  "paths": {
...
    "/auth/token": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Issue JWT (demo)",
        "operationId": "auth_token_auth_token_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/ingest": {
      "post": {
        "tags": [
          "telemetry"
        ],
        "summary": "Direct ingest of a normalized reading (for gateways)",
        "operationId": "ingest_reading_ingest_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestReadingIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IngestReadingIn": {
        "properties": {
          "station_id": {
            "type": "string",
            "title": "Station Id"
          },
          "metric": {
            "type": "string",
            "title": "Metric"
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "unit": {
            "type": "string",
            "title": "Unit",
            "default": ""
          },
          "protocol": {
            "type": "string",
            "title": "Protocol",
            "default": "unknown"
          },
          "adapter_id": {
            "type": "string",
            "title": "Adapter Id",
            "default": ""
          },
          "source_tag": {
            "type": "string",
            "title": "Source Tag",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "station_id",
          "metric",
          "value"
        ],
        "title": "IngestReadingIn"
      },
...
      "TokenIn": {
        "properties": {
          "sub": {
            "type": "string",
            "title": "Sub",
            "default": "operator"
          },
          "plant_id": {
            "type": "string",
            "title": "Plant Id",
            "default": "plant-demo-01"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "default": "engineer"
          },
          "exp_min": {
            "type": "integer",
            "title": "Exp Min",
            "default": 60
          }
        },
        "type": "object",
        "title": "TokenIn"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}